now it should work

+++ internal/application/services/comment_service.go
@@ -4,7 +4,7 @@ import "github.com/niflheimdevs/backend/internal/application/dto"
 
 type CommentService interface {
 	PutComment(userID, ProjectID int, content string, star int) int
-	GetUserStar(userID int) float32
+	GetUserStar(userID int) float64
 	GetUserComments(userID int) []dto.CommentDTO
 	GetCommentInfo(id int) dto.CommentDTO
 }
